Create users_password integration test
